moved some styles directly to the Box component as props and refactored all the files in helpers/ManageBookings/ into a single file in helpers/ManageBookings.tsx
