--- client/src/helpers/ManageBookings/index.ts
@@ -1,15 +1,0 @@
-import isThereBookingOnThisDate from './isThereBookingOnThisDate';
-import formatBookingDate from './formatBookingDate';
-import sortBookingsByStartDate from './sortBookingsByStartDate';
-import dateIsTodayOrAfter from './dateIsTodayOrAfter';
-import getCurrentBookings from './getCurrentBookings';
-import getNextBooking from './getNextBooking';
-
-export {
-  isThereBookingOnThisDate,
-  formatBookingDate,
-  sortBookingsByStartDate,
-  dateIsTodayOrAfter,
-  getCurrentBookings,
-  getNextBooking,
-};
